word2vec_from_scratch/requirements.txt
datasets==3.6.0
nltk==3.9.1
numpy==2.3.1
pandas==2.3.0
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.0
streamlit==1.46.1
torch==2.7.1+cpu
